Experiment show-casing default hashing function
Our hashing function (that we use to split users into test buckets based
on keys that might be similar to GUIDs) is pretty simplistic. But the
goal of uniform distribution may or may not work well.

In this experiment, we create thousands of GUIDs and use our hashing
function as well as a Murmur Hash function and graphically show how they
compare.